Move test of _is_webdav_endpoint() to test_http.py and improve testing of _do_put()
